requirements.txt
chatterbox-tts>=0.1.6
soundfile>=0.12.0
python-dotenv>=1.0.0
openai>=2.14.0
numpy<1.26.0,>=1.24.0